--- spring-boot/src/main/java/com/example/spring_boot/adapters/RepoAdapter.java
@@ -1,13 +1,0 @@
-package com.example.spring_boot.adapters;
-
-import com.example.spring_boot.models.EventDataRecord;
-
-import java.util.stream.Stream;
-
-public interface RepoAdapter {
-
-    public String getDbName();
-
-    public Stream<EventDataRecord> processPartition(long id,long start,long end);
-
-}
