--- spring-boot/src/main/java/com/example/spring_boot/dto/EventDataPayload.java
@@ -1,4 +1,0 @@
-package com.example.spring_boot.dto;
-
-public record EventDataPayload(String db,long deviceId,long startTime,long endTime,int page,int size,boolean isDaily) {
-}
