--- spring-boot/src/main/java/com/example/spring_boot/models/EventDataRecord.java
@@ -1,4 +1,0 @@
-package com.example.spring_boot.models;
-
-public record EventDataRecord(double latitude, double longitude, long timestamp) {
-}
