+++ src/main/java/com/example/eventdata_service/models/EventDataRecord.java
@@ -1,0 +1,4 @@
+package com.example.eventdata_service.models;
+
+public record EventDataRecord(double latitude, double longitude, long timestamp) {
+}
