+++ src/main/java/com/example/eventdata_service/models/TackPayload.java
@@ -1,0 +1,14 @@
+package com.example.eventdata_service.models;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+import java.sql.Timestamp;
+
+@Data
+@AllArgsConstructor
+public class TackPayload {
+    long startTime;
+    long endTime;
+    boolean isDailyMode;
+}
